==> 01_worldcup.ipynb <==
import pandas as pd
import pickle
from string import ascii_uppercase as alphabet

# Scrap tables from web
all_tables = pd.read_html("https://web.archive.org/web/20221115040351/https://en.wikipedia.org/wiki/2022_FIFA_World_Cup")

# Initialize dictionary
dict_table = {}

# Loop through each table of interest
for letter, i in zip(alphabet, range(12, 68, 7)):
    # Create Dataframe for each group in World Cup
    df = all_tables[i]
    # Rename the column 'Team'
    df.rename(columns={df.columns[1]: "Team"}, inplace=True)
    # Delete the 'Qualification' column
    df.pop("Qualification")
    # Populate dictionary
    dict_table[f"Group {letter}"] = df

# Inspect Dictionary
dict_table["Group A"]

# Save dictionary to binary file using pickle
with open("dict_table.pkl", "wb") as output:
    pickle.dump(dict_table, output)
